refactor node image fns into new fild and create fn for creating and deleteing apps (folders)
