--- applications/test1/hi.txt
@@ -1,1 +1,0 @@
-hello
